EE-904: address @EdHastingsCasperLabs review comment
