+++ execution-engine/engine-tests/src/test/deploy/mod.rs
@@ -1,2 +1,3 @@
+mod non_standard_payment;
 mod preconditions;
 mod stored_contracts;
